% Legacy documentation retained for historical analyses.
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_Aldex2_V1.R
\name{Go_Aldex2}
\alias{Go_Aldex2}
\title{Perform Differential Abundance Analysis Using ALDEx2 on Phyloseq Data}
\usage{
Go_Aldex2(
  psIN,
  project,
  cate.outs = NULL,
  cont.outs = NULL,
  cate.conf = NULL,
  cont.conf = NULL,
  orders = NULL,
  name = NULL
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{project}{Name of the project or analysis.}

\item{cate.outs}{Categorical outcomes to be analyzed.}

\item{cont.outs}{Continuous outcomes to be analyzed.}

\item{cate.conf}{Categorical confounding variables to adjust for in the analysis.}

\item{cont.conf}{Continuous confounding variables to adjust for in the analysis.}

\item{orders}{Custom order of factors in the analysis, if applicable.}

\item{name}{Optional name for the analysis.}
}
\value{
The function generates CSV files containing the results of the ALDEx2 differential abundance analysis. These include statistics like effect sizes, p-values, and adjusted p-values, along with taxonomic information for each feature. Files are saved in a specified directory with a naming convention that includes key details of the analysis.
}
\description{
This function conducts differential abundance analysis using the ALDEx2 method on data from a Phyloseq object. It supports various types of analyses including categorical and continuous variables, and allows for adjusting for confounders. The function can handle multiple categories, and generate output for each pairwise comparison.
}
\details{
The function first checks the type of data (taxonomy, kegg, pathway, RNAseq) in the Phyloseq object. It then processes categorical and continuous outcomes, adjusting for confounders if specified. ALDEx2 analysis is performed for each variable or combination of variables. The results are merged with taxonomy data and saved as CSV files.

For categorical variables, the function considers combinations of levels for analysis. For continuous variables, it performs correlation analysis. The function allows for multiple comparisons and adjusts the results using methods like GLM or t-test based on the type of variable.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_Aldex2(psIN = psIN,
          project = "MyProject",
          cate.outs = c("Treatment", "Condition"),
          cont.outs = c("Age", "BMI"),
          cate.conf = c("Gender"),
          cont.conf = NULL,
          orders = NULL,
          name = "Analysis1")

}
