% Legacy documentation retained for historical analyses.
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Go_Deseq2_v1.R
\name{Go_Deseq2}
\alias{Go_Deseq2}
\title{Perform Differential Expression Analysis Using DESeq2 on Phyloseq Data}
\usage{
Go_Deseq2(
  psIN,
  project,
  taxanames = NULL,
  cate.outs,
  cate.conf = NULL,
  cont.conf = NULL,
  orders = NULL,
  name = NULL
)
}
\arguments{
\item{psIN}{Phyloseq object containing the data for analysis.}

\item{project}{Name of the project or analysis.}

\item{taxanames}{Taxonomic names to aggregate the data. NULL if no aggregation is needed.}

\item{cate.outs}{Categorical outcomes for differential expression analysis.}

\item{cate.conf}{Categorical confounding variables to adjust for in the analysis.}

\item{cont.conf}{Continuous confounding variables to adjust for in the analysis.}

\item{orders}{Custom order of factors in the analysis, if applicable.}

\item{name}{Optional name for the analysis.}
}
\value{
Generates CSV files containing the results of the DESeq2 analysis,
including statistics like log fold change, p-values, and adjusted p-values,
along with taxonomic information for each feature. Files are saved in a specified directory.
}
\description{
This function conducts differential expression analysis using DESeq2 on data from a Phyloseq object.
It allows for the analysis of categorical outcomes, with the option to control for categorical
and continuous confounding variables. The function can handle data aggregated at different taxonomic levels.
}
\details{
The function preprocesses the Phyloseq object, potentially aggregating data at a specified taxonomic level.
It then performs DESeq2 analysis for each categorical outcome, adjusting for confounders if specified.
The results are merged with taxonomy data and saved as CSV files. The function supports pairwise comparisons
within categorical outcomes and handles data with or without specified confounders.
}
\examples{
# psIN is a Phyloseq object
# Example usage:
Go_Deseq2(psIN = psIN,
          project = "MyProject",
          taxanames = "Genus",
          cate.outs = c("Treatment", "Condition"),
          cate.conf = c("Gender"),
          cont.conf = NULL,
          orders = NULL,
          name = "Analysis1")

}
