% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegacyNNS_Distance.R
\name{LegacyNNS.distance}
\alias{LegacyNNS.distance}
\title{LegacyNNS Distance}
\usage{
LegacyNNS.distance(rpm, dist.estimate, k, class)
}
\arguments{
\item{rpm}{REGRESSION.POINT.MATRIX from \link{LegacyNNS.reg}}

\item{dist.estimate}{Vector to generate distances from.}

\item{k}{\code{n.best} from \link{LegacyNNS.reg}}

\item{class}{if classification problem.}
}
\value{
Returns sum of weighted distances.
}
\description{
Internal kernel function for LegacyNNS multivariate regression \link{LegacyNNS.reg} parallel instances.
}
