% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{LegacyNNS.TSD.uni}
\alias{LegacyNNS.TSD.uni}
\title{LegacyNNS TSD Test uni-directional}
\usage{
LegacyNNS.TSD.uni(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
Returns (1) if \code{"X TSD Y"}, else (0).
}
\description{
Uni-directional test of third degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
\dontrun{
set.seed(123)
x <- rnorm(100) ; y <- rnorm(100)
LegacyNNS.TSD.uni(x, y)
}
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126.  \doi{10.4236/jmf.2016.61012}.
}
\author{
Fred Viole, OVVO Financial Systems
}
